chore: rename cuewrapper to openapi
